Add support for combined fields in the regular dataform view as well
